fix: change in migrations

+++ database/migrations/2025_05_28_075830_create_products_table.php
@@ -24,7 +24,7 @@ return new class extends Migration {
             $table->string('name');
             $table->integer('category_id');
             $table->text('description');
-            $table->integer('price');
+            $table->decimal('price', 10);
             $table->timestamps();
         });
     }
